some changes

+++ lib/view/screens/my_membership_screen/membership_profile_model/membership_profile_model.dart
@@ -115,7 +115,7 @@ class Plan {
 class PlanId {
   String? id;
   String? planName;
-  double? fee;
+  int? fee;
   int? pointRangeStart;
   int? pointRangeEnd;
   int? swapPoint;
